Fix warning detected by msvc 6
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@15886 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/blast/core/blast_util.c
@@ -969,7 +969,7 @@ int GetPartialTranslation(const Uint1* nucl_seq,
       for (index = 1; index <= 3; ++index) {
          length = 
             BLAST_GetTranslation(nucl_seq, nucl_seq_rev, 
-               nucl_length, frame_sign*index, translation_buffer+offset, 
+               nucl_length, (short)(frame_sign*index), translation_buffer+offset, 
                genetic_code);
          frame_offsets[index-1] = offset;
          offset += length + 1;
